+++ ErsatzTV.Application/Resolutions/Mapper.cs
@@ -5,5 +5,5 @@ namespace ErsatzTV.Application.Resolutions;
 internal static class Mapper
 {
     internal static ResolutionViewModel ProjectToViewModel(Resolution resolution) =>
-        new(resolution.Id, resolution.Name, resolution.Width, resolution.Height);
+        new(resolution.Id, resolution.Name, resolution.Width, resolution.Height, resolution.IsCustom);
 }