+++ ErsatzTV.Infrastructure/Data/Configurations/ResolutionConfiguration.cs
@@ -6,5 +6,10 @@ namespace ErsatzTV.Infrastructure.Data.Configurations;
 
 public class ResolutionConfiguration : IEntityTypeConfiguration<Resolution>
 {
-    public void Configure(EntityTypeBuilder<Resolution> builder) => builder.ToTable("Resolution");
+    public void Configure(EntityTypeBuilder<Resolution> builder)
+    {
+        builder.ToTable("Resolution");
+
+        builder.Property(r => r.IsCustom).HasDefaultValue(false);
+    }
 }