+++ ErsatzTV/ViewModels/ResolutionEditViewModel.cs
@@ -1,0 +1,7 @@
+namespace ErsatzTV.ViewModels;
+
+public class ResolutionEditViewModel
+{
+    public int Width { get; set; }
+    public int Height { get; set; }
+}